requirements.txt
alembic==1.10.3
blinker==1.6
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==0.21.1
SQLAlchemy==2.0.9
typing-extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0